Merged PR 599: Merge v-jamedr/MosaicPositioning to master
I have added a few things, want to get your thoughts.

add alignX to header to switch left, right, or center.

add pictureblock to header to show an image in the header

add alignY to pictureBlock to position the picture

immersiveHero now needs a pictureblock in the header to show the foreground image. the other image will be used in the background. alignY will still position the foreground image above or below the text block.

You can test out these changes in Storybook.